% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graw.R
\name{addlabels}
\alias{addlabels}
\title{Add text label on board}
\usage{
addlabels(gg, x, y, label, color = NULL, ...)
}
\arguments{
\item{gg}{\code{ggplot} object}

\item{x, y}{integer vectors of stone locations}

\item{label}{character vector of labels}

\item{color}{integer vector of stone colors}

\item{...}{graphic paramters}
}
\value{
Updated \code{ggplot} object
}
\description{
Add text label on board
}
\examples{
ggoban(19) \%>\% addstones(c(16, 4), c(16, 3), c(1, 2)) \%>\%
  addlabels(4, 3, "X", 2) \%>\%
  addlabels(c(4, 3), c(17, 16), c("a", "b"))
}

