% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw.R
\name{addmarkers}
\alias{addmarkers}
\title{Add markers}
\usage{
addmarkers(gg, x, y, color, marker = 17, ...)
}
\arguments{
\item{gg}{\code{ggplot} object}

\item{x, y}{integer vectors of stone locations}

\item{color}{integer vector of stone colors}

\item{marker}{scalar integer indicating the shape of marker}

\item{...}{graphic paramters}
}
\value{
Updated \code{ggplot} object
}
\description{
Add markers
}

