% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw.R
\name{addstones}
\alias{addstones}
\title{Add stones to go board}
\usage{
addstones(gg, x, y, color, number = NULL, ...)
}
\arguments{
\item{gg}{\code{ggplot} object}

\item{x, y}{integer vectors of stone locations}

\item{color}{integer vector of stone colors}

\item{number}{integer vector of numbers on stones}

\item{...}{graphic paramters}
}
\value{
Updated \code{ggplot} object
}
\description{
Add stones to go board
}
\examples{
ggoban(19) \%>\% addstones(c(10, 11), c(10, 10), c(1, 2), c(10, 11))
ggoban(9) \%>\%
 addstones(x = c(5, 5, 4), y = c(5, 3, 3),
           color = c(1, 2, 1), boardsize = 9)
}

