% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw.R
\name{addterritory}
\alias{addterritory}
\title{Add territory markers}
\usage{
addterritory(gg, x, y, color, ...)
}
\arguments{
\item{gg}{\code{ggplot} object}

\item{x, y}{integer vectors of stone locations}

\item{color}{integer vector of stone colors}

\item{...}{graphic paramters}
}
\value{
Updated \code{ggplot} object
}
\description{
Add territory markers
}
\seealso{
\code{\link{ggoban}}, \code{\link{graphic_parameters}}
}

