% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{get_points}
\alias{get_points}
\title{Obtains the point markers in sgf}
\usage{
get_points(sgf)
}
\arguments{
\item{sgf}{Scalar character of sgf text}
}
\value{
\code{data.frame} with variables x, y, and color indicating
the location of points on the board
}
\description{
Obtains the point markers in sgf
}

