% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tags.R
\name{get_props}
\alias{get_props}
\title{Find the tag properties in sgf text.}
\usage{
get_props(sgf, tags)
}
\arguments{
\item{sgf}{A character string of sgf format}

\item{tags}{A character vector of tags}
}
\value{
A character vector of the same size as \code{tags}.
  It contains the properties of tags (\code{NA} if the tag is missing)
  and is named by the tag names.
}
\description{
Find the tag properties in sgf text.
}
\details{
This function finds the first appearance of each tag
  and returns the property (contents within the bracket).
  It suits for finding propeties unique to a game
  (e.g. meta information, date, players, rule set),
  but not for extracting tags appearing for multiple times, such as moves.
}
\examples{
sgf <- "(;GM[1]PW[Iyama]PB[Yamashita]RO[Final]EV[Kisei])"
find_tags(sgf, "PB")
find_tags(sgf, c("PW", "PB", "EV", "RO"))

}

