% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transition.R
\name{get_transition}
\alias{get_transition}
\title{Obtain the board transition from moves}
\usage{
get_transition(moves)
}
\arguments{
\item{moves}{Moves}
}
\value{
transition.
}
\description{
Obtain the board transition from moves
}

