% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_transitions}
\alias{get_transitions}
\title{Obtains the transition of board configuration}
\usage{
get_transitions(boardsize, ismove_vec, x_vec, y_vec, color_vec)
}
\value{
\code{data.frame}
}
\description{
Obtains the transition of board configuration
}

