% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_transitiontree}
\alias{get_transitiontree}
\title{Obtains the transition of board configuration}
\usage{
get_transitiontree(data, children, boardsize)
}
\arguments{
\item{data}{A list of data.frames, each contains}
}
\description{
Obtains the transition of board configuration
}
\keyword{internal}

