% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggoban-class.R
\name{ggoban}
\alias{ggoban}
\title{Draw go board}
\usage{
ggoban(boardsize, ...)
}
\arguments{
\item{boardsize}{integer of boardsize}

\item{...}{graphic parameters}
}
\value{
Object of class \code{ggoban}, which inherits \code{ggplot}
}
\description{
Draw go board
}
\examples{
ggoban(19)
}

