% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gogame-class.R
\name{ggoboard}
\alias{ggoboard}
\title{ggplot go board}
\usage{
ggoboard(boardsize, gridcolor = "#262626", boardcolor = "#e1f0c0")
}
\arguments{
\item{boardsize}{integer of boardsize}

\item{gridcolor}{color for the grid}

\item{boardcolor}{color for the background}
}
\value{
\code{ggplot} object of goban
}
\description{
ggplot go board
}
\examples{
  ggoboard(19)
}

