% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gogame-class.R
\name{gogame}
\alias{gogame}
\alias{is.gogame}
\title{Go game object}
\usage{
gogame(properties, moves)

is.gogame(x)
}
\arguments{
\item{properties}{a list of game properties}

\item{moves}{a data frame of game moves}

\item{x}{An R object}
}
\value{
\code{gogame} object

Logical.
}
\description{
\code{gogame} stores go game record including
game plays, results and
other information such as player names and game setup.
}

