% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gogamer.R
\docType{package}
\name{gogamer}
\alias{gogamer}
\alias{gogamer-package}
\title{gogamer: Go Game Data Parser}
\description{
Easy and flexible interface for manipulating go game (weiqi, baduk) data.
The package features a reader function for SGF (smart go format) text files,
and a set of plotting functions that draw go board images.
}

