% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggoban-class.R
\name{is.ggoban}
\alias{is.ggoban}
\title{Check if object is ggoban class}
\usage{
is.ggoban(obj)
}
\arguments{
\item{obj}{R object}
}
\value{
Logical. True if and only if \code{obj} inherits \code{ggoban} class.
}
\description{
Check if object is ggoban class
}

