% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gostate-class.R
\name{is.gostate}
\alias{is.gostate}
\title{Check if the object is gostate class}
\usage{
is.gostate(x)
}
\arguments{
\item{x}{R object}
}
\value{
logical
}
\description{
Check if the object is gostate class
}

