% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kifu-class.R
\name{kifu.plot}
\alias{kifu.plot}
\title{Draw kifu}
\usage{
kifu.plot(obj, stonesize = 6, blackcolor = "#000000",
  whitecolor = "#ffffff", edgecolor = "#000000", textcolor = "#000000")
}
\arguments{
\item{obj}{\code{kifu} object}
}
\value{
\code{ggplot} object
}
\description{
Draw kifu
}

