% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kifu-class.R
\name{kifu}
\alias{kifu}
\title{Generate kifu (game record) object}
\usage{
kifu(x, from = 1L, to = 100L)
}
\arguments{
\item{x}{\code{gogame} object}

\item{from, to}{integers specifying the range of moves`}
}
\value{
\code{kifu} object
}
\description{
Generate kifu (game record) object
}

