% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kifuplot-class.R
\name{kifuimage}
\alias{kifuimage}
\title{Constructor of kifuplot object}
\usage{
kifuimage(board, note, boardsize, notenrow)
}
\arguments{
\item{board}{\code{ggplot} object of board}

\item{note}{\code{ggplot} object of outside note}

\item{boardsize}{integer of board size}

\item{notenrow}{integer of number of rows of note}
}
\value{
\code{kifuimage} object
}
\description{
Constructor of kifuplot object
}

