% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kifu-class.R
\name{kifunote}
\alias{kifunote}
\title{Draw outside note of kifu}
\usage{
kifunote(x, ...)
}
\arguments{
\item{x}{\code{kifu} object}

\item{...}{graphic parameter}
}
\value{
\code{ggplot} object
}
\description{
Draw outside note of kifu
}

