% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mimiaka}
\alias{mimiaka}
\title{Mimiaka game}
\format{\code{gogame} object}
\source{
TBA
}
\usage{
mimiaka
}
\description{
Go game known as "Mimiaka game", between Genan Inseki
and Kuwabara Shusaku
}
\details{
TBA
}
\keyword{datasets}

