% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sgfreader.R
\name{parse_sgf}
\alias{parse_sgf}
\title{Parse text of the smart go format.}
\usage{
parse_sgf(sgf)
}
\arguments{
\item{sgf}{scalar character of sgf text.}
}
\value{
\code{\link{gogame}} object
}
\description{
Parse text of the smart go format.
}

