% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sgftagparser.R
\name{parse_sgfnode}
\alias{parse_sgfnode}
\title{Parse SGF of each node}
\usage{
parse_sgfnode(sgf)
}
\arguments{
\item{sgf}{A character vector of SGF nodes}

\item{asList}{if this is True, each component is list of the node size}
}
\value{
A list
}
\description{
Parse SGF of each node
}
\keyword{internal}

