% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-and-parse.R
\name{parseSGF}
\alias{parseSGF}
\title{Parse text in the smart go format.}
\usage{
parseSGF(sgf)
}
\arguments{
\item{sgf}{A character vector of sgf text.}
}
\description{
Parse text in the smart go format.
}

