% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kifu-class.R
\name{plot.kifu}
\alias{plot.kifu}
\title{Draw kifu}
\usage{
\method{plot}{kifu}(x, y, ...)
}
\arguments{
\item{x}{\code{kifu} object}

\item{y}{not in use (just for argument consistency with generic function)}

\item{...}{graphical paramters}
}
\value{
list of two \code{ggplot} object,
one for the board, the other for outside note
}
\description{
Draw kifu
}

