% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gogame-class.R
\name{plotat}
\alias{plotat}
\title{Plot the go board state by ggplot}
\usage{
plotat(x, at, ...)
}
\arguments{
\item{x}{\code{gogame} object}

\item{at}{Move number (integer)}

\item{...}{arguments passed to \code{\link{print.gostate}}}
}
\value{
\code{ggplot} object
}
\description{
Plot the go board state by ggplot
}
\examples{
plotat(mimiaka, 127)
}

