% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggkifu-class.R
\name{print.ggkifu}
\alias{print.ggkifu}
\title{Print kifu on concole}
\usage{
\method{print}{ggkifu}(x, quiet = FALSE, ...)
}
\arguments{
\item{x}{\code{ggkifu} object}

\item{quiet}{Logical. If true, print suggested size for saving}

\item{...}{Not in use}
}
\description{
Print kifu on concole
}

