% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gokifu-class.R
\name{print.gokifu}
\alias{print.gokifu}
\title{Print kifu on console}
\usage{
\method{print}{gokifu}(x, adjust_origin = TRUE, ...)
}
\arguments{
\item{x}{\code{gokifu} object}

\item{adjust_origin}{Logical. If this is true, numbers are
deducted by a multiple of 100 if appropriate}

\item{...}{graphical parameters}
}
\description{
Print kifu on console
}

