% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gostate-class.R
\name{print.gostate}
\alias{print.gostate}
\title{Draw board state on console}
\usage{
\method{print}{gostate}(x, ...)
}
\arguments{
\item{x}{\code{gostate} object}

\item{...}{graphic parameters}
}
\description{
Draw board state on console
}
\examples{
stateat(mimiaka, 127)
}

