% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-and-parse.R
\name{readSGF}
\alias{readSGF}
\title{Read and parse a SGF file}
\usage{
readSGF(path, encoding = "")
}
\arguments{
\item{path}{A character representing the location of a smart go format (SGF) file.}

\item{encoding}{A character string: if non-empty declares the encoding used on a file.}
}
\description{
Read and parse a SGF file
}

