% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{saikoyo}
\alias{saikoyo}
\title{sai vs toya koyo}
\format{\code{gogame} object}
\source{
TBA
}
\usage{
saikoyo
}
\description{
Go game between sai and toya koyo played online
in the comic "Hikaru no Go"
}
\details{

}
\keyword{datasets}

