% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gogame-class.R
\name{set_branch}
\alias{set_branch}
\title{Switch branch of go game}
\usage{
set_branch(x, branch = 1L)
}
\arguments{
\item{x}{\code{gogame} object}

\item{branch}{integer}
}
\value{
\code{gogame} object
}
\description{
Switch branch of go game
}

