% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gogame-class.R
\name{set_gamepath}
\alias{set_gamepath}
\title{Switch game path of go game}
\usage{
set_gamepath(x, pathid = 1L)
}
\arguments{
\item{x}{\code{gogame} object}

\item{branch}{integer}
}
\value{
\code{gogame} object
}
\description{
Switch game path of go game
}

