% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goboard.R
\name{set_graphic_param}
\alias{set_graphic_param}
\title{Set graphic parameters for go board}
\usage{
set_graphic_param(...)
}
\arguments{
\item{...}{Graphic paramters}
}
\value{
list of graphic parameters
}
\description{
Set graphic parameters for go board
}
\details{
this function set the graphic parameters of goban image
uses default (.default_graph_param) in principle
replace the value specified in '...'
default values are not modified
returns a list of graphic parameters
}

