% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sgf.R
\docType{package}
\name{sgf}
\alias{sgf}
\alias{sgf-package}
\title{sgf: A package for parsing a SGF (smart go format) file.}
\description{
This package provides utilities for parsing text files in smart go format.
}
\section{Reader}{

\code{readSGF}
}

