% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goboard.R
\name{star_position}
\alias{star_position}
\title{Get location of stars}
\usage{
star_position(boardsize)
}
\arguments{
\item{boardsize}{integer of boardsize}
}
\value{
\code{data.frame} with variables \code{x} and \code{y}
}
\description{
Get location of stars
}

