% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gogame-class.R
\name{stateat}
\alias{stateat}
\title{Return the board state}
\usage{
stateat(x, at)
}
\arguments{
\item{x}{\code{gogame} object}

\item{at}{integer of the move number}
}
\value{
data frame object
}
\description{
Return the board state
}

