% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggkifu-class.R
\name{suggested_size}
\alias{suggested_size}
\title{Suggested size used for saving}
\usage{
suggested_size(obj)
}
\arguments{
\item{obj}{\code{ggkifu} object}
}
\value{
Numeric vector of size two
}
\description{
Suggested size used for saving
}

