% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tsumego}
\alias{tsumego}
\title{Tsumego (Life and death problem)}
\format{\code{gogame} object}
\source{
Cho U (Zhang Xu). Kihon Shikatsu Jiten. Nihon Kiin 2014.
}
\usage{
tsumego
}
\description{
A basic life and death problem. It shows three variations,
the first is failure, and the last two are the success cases.
}
\keyword{datasets}

