% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphic.R
\name{update_graphic_param}
\alias{update_graphic_param}
\title{Update graphic parameters}
\usage{
update_graphic_param(params, ...)
}
\arguments{
\item{params}{A list of graphic parameters}

\item{...}{Graphic parameters}
}
\value{
An updated list of graphic parameters
}
\description{
Update graphic parameters
}

