% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goboard.R
\name{addstones}
\alias{addstones}
\title{Add stones to go board}
\usage{
addstones(x, y, color, label = NULL)
}
\arguments{
\item{x, y}{integer vectors of stone locations}

\item{color}{integer vector of stone colors}

\item{label}{character vector of labels added on the stone}
}
\value{
Layer for adding to \code{ggplot}
}
\description{
Add stones to go board
}

