% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tags.R
\name{get_moves}
\alias{get_moves}
\title{Obtain game sequences (plays, comments, and times) in sgf text.}
\usage{
get_moves(sgf)
}
\arguments{
\item{sgf}{a character of text in sgf format}
}
\value{
an integer matrix with five columns.
}
\description{
Obtain game sequences (plays, comments, and times) in sgf text.
}

