% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{get_moves}
\alias{get_moves}
\title{Obtain setup and plays in sgf text}
\usage{
get_moves(sgf)
}
\arguments{
\item{sgf}{scalar character of text in sgf format}
}
\value{
a \code{data.frame} object with four columns:
  \describe{
    \item{color}{color (1: black, 2: white)}
    \item{x}{x coordinate}
    \item{y}{y coordinate}
    \item{ismove}{TRUE: move, FALSE: setup}
  }
}
\description{
Obtain setup and plays in sgf text
}
\examples{
get_moves("AB[pd][dp][pp][dd]W[dg]")
}

