% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sgfparser.R
\name{get_points}
\alias{get_points}
\title{Obtains the point markers in sgf}
\usage{
get_points(sgf)
}
\arguments{
\item{sgf}{Scalar character of sgf text}
}
\value{
\code{data.frame} with variables \code{x}, \code{y}, and \code{color}
}
\description{
Obtains the point markers in sgf
}

