% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sgftagparser.R
\name{get_props}
\alias{get_props}
\title{Find tag properties in sgf}
\usage{
get_props(sgf, tags)
}
\arguments{
\item{sgf}{scalar character of texts in sgf format}

\item{tags}{character vector of tags}
}
\value{
Named list of game properties
}
\description{
Find tag properties in sgf
}
\details{
This function finds the first appearance of each tag
  and returns the property (contents within the bracket).
  It suits for finding propeties unique to a game
  (e.g. player names or rule set),
  but not for extracting tags appearing for multiple times, such as moves.
}
\keyword{internal}

