% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggkifu-class.R
\name{ggkifu}
\alias{ggkifu}
\title{One-page kifu image}
\usage{
ggkifu(board, note, boardsize, heights = NULL, savesize = NULL)
}
\arguments{
\item{board}{\code{ggplot} object of board}

\item{note}{\code{ggplot} object of outside note}

\item{boardsize}{integer of board size}

\item{heights}{Numeric vector of size two, indicates the vertical size
ratio between board and note}

\item{savesize}{Numeric vector of size two, indicating
appropriate pair of width and height when saving}
}
\value{
\code{ggkifu} object
}
\description{
One-page kifu image
}

