% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggkifu-class.R
\name{ggkifu}
\alias{ggkifu}
\title{One-page kifu image}
\usage{
ggkifu(board, note, boardsize, notenrow)
}
\arguments{
\item{board}{\code{ggplot} object of board}

\item{note}{\code{ggplot} object of outside note}

\item{boardsize}{integer of board size}

\item{notenrow}{integer of number of rows of note}
}
\value{
\code{ggkifu} object
}
\description{
One-page kifu image
}

