% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gogame-class.R
\name{gogame}
\alias{gogame}
\title{Constructor of gogame object}
\usage{
gogame(properties, moves)
}
\arguments{
\item{properties}{a list of game properties}

\item{moves}{a data frame of game moves}
}
\value{
\code{gogame} object
}
\description{
Constructor of gogame object
}

