% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gogamer.R
\docType{package}
\name{gogamer}
\alias{gogamer}
\alias{gogamer-package}
\title{gogamer: A package for parsing a SGF (smart go format) file.}
\description{
This package provides utilities for parsing text files in smart go format.
}
\section{Reader}{

\code{readSGF}
}

