% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gokifu-class.R
\name{gokifu}
\alias{gokifu}
\title{Kifu (go game record) for a range of moves}
\usage{
gokifu(init, numbered, noted, boardsize)
}
\arguments{
\item{init}{initial board state (\code{data.frame})}

\item{numbered}{moves to be numberd on the board (\code{data.frame})}

\item{noted}{moves to be listed ontside the board (\code{data.frame})}

\item{boardsize}{integer of board size}
}
\value{
\code{gokifu} object
}
\description{
Kifu (go game record) for a range of moves
}

