% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gokifu-class.R
\name{gokifu}
\alias{gokifu}
\title{Kifu (go game record) for a range of moves}
\usage{
gokifu(unnumbered, numbered, noted, comment, boardsize)
}
\arguments{
\item{unnumbered}{stones to be shown with no number on the board.
These typically are stones put before the move range of the kifu.
It also includes setup stones in the range (\code{data.frame})}

\item{numbered}{moves to be numberd on the board (\code{data.frame})}

\item{noted}{moves to be listed ontside the board (\code{data.frame})}

\item{comment}{comments during the move range of the game (\code{data.frame})}

\item{boardsize}{board size (\code{integer})}
}
\value{
\code{gokifu} object
}
\description{
Kifu (go game record) for a range of moves
}

