% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gostate-class.R
\name{gostate}
\alias{gostate}
\title{Go game state}
\usage{
gostate(board, boardsize, b_captured, w_captured, movenumber = NULL,
  lastmove = NULL, points = NULL, comment = NULL)
}
\arguments{
\item{board}{a \code{data.frame} representing stone allocation.
It must have variables (\code{x}, \code{y}, \code{color})}

\item{boardsize}{baord size (integer)}

\item{b_captured, w_captured}{numbers of captured stone (integer)}

\item{movenumber}{integer of move number}

\item{lastmove}{integer vector of length three that indicates the last move
location and color in the order of (x, y, color)}

\item{points}{\code{data.frame} of territory locations}

\item{comment}{character vectror of comments}
}
\value{
\code{gostate} object
}
\description{
\code{gostate} object stores a go game state including
stone configuration on the board and the numbers of prisoners.
}
\examples{
gostate(data.frame(x = 4, y = 4, color = 1), 19, 0, 0)
}

