% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gogame-class.R
\name{kifu}
\alias{kifu}
\title{Kifu for certain move range}
\usage{
kifu(x, from = 1L, to = 100L, restart = NA_integer_)
}
\arguments{
\item{x}{\code{gogame} object}

\item{from, to}{Positive integers. Range of moves}

\item{restart}{Positive integer. If supplied, this number is used as the
smallest move number in the range. If not supplied, original move numbers
are used as they are.}
}
\value{
\code{\link{gokifu}} object
}
\description{
Kifu for certain move range
}
\examples{
kifu(saikoyo)
}

