% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gogame-class.R
\name{kifu}
\alias{kifu}
\title{Kifu for certain move range}
\usage{
kifu(x, from = 1L, to = 100L)
}
\arguments{
\item{x}{\code{gogame} object}

\item{from, to}{integers specifying range of move}
}
\value{
\code{\link{gokifu}} object
}
\description{
Kifu for certain move range
}

