% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mimiaka}
\alias{mimiaka}
\title{Mimiaka (ear-reddening) game}
\format{\code{gogame} object}
\source{
\itemize{
  \item{Power, John. Invincible: The games of Shusaku. Kiseido Publishing Company 1982.}
  \item{\url{http://senseis.xmp.net/?EarReddeningMove}}
}
}
\usage{
mimiaka
}
\description{
Go game known as "Mimiaka (ear-redding) game", between Genan Inseki
and Kuwabara Shusaku
}
\keyword{datasets}

