% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reader.R
\name{parse_sgf}
\alias{parse_sgf}
\title{Parse text of the smart go format.}
\usage{
parse_sgf(sgf, keepfirst = TRUE)
}
\arguments{
\item{sgf}{scalar character of sgf text.}

\item{keepfirst}{logical specifying the branch choice rule.
If TRUE, keep the first appearing branch.
Otherwise, keep the last branch.}
}
\value{
\code{\link{gogame}} object
}
\description{
Parse text of the smart go format.
}

