% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gokifu-class.R
\name{plot.gokifu}
\alias{plot.gokifu}
\title{Draw kifu}
\usage{
\method{plot}{gokifu}(x, y, ...)
}
\arguments{
\item{x}{\code{kifu} object}

\item{y}{not in use (just for argument consistency with generic function)}

\item{...}{graphical paramters}
}
\value{
\code{\link{ggkifu}} object
}
\description{
Draw kifu
}

