% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gostate-class.R
\name{plot.gostate}
\alias{plot.gostate}
\title{Draw go board state as graphic}
\usage{
\method{plot}{gostate}(x, y, marklast = TRUE, markpoints = FALSE, ...)
}
\arguments{
\item{x}{\code{gostate} object}

\item{y}{not in use}

\item{marklast}{logical indicating if last move should be marked}

\item{markpoints}{logical indicating if territories should be marked}

\item{...}{graphic parameters}
}
\value{
\code{ggoban} object, which inherits \code{ggplot} class
}
\description{
Draw go board state as graphic
}
\examples{
stateat(saikoyo, 116) \%>\% plot()
}
\seealso{
\code{\link{gogame_graphics}}
}

