% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kifu-class.R
\name{plot.kifu}
\alias{plot.kifu}
\title{Draw kifu}
\usage{
\method{plot}{kifu}(x, y, stonesize = 6, blackcolor = "#000000",
  whitecolor = "#ffffff", linecolor = "#000000", textcolor = "#000000",
  ...)
}
\arguments{
\item{x}{\code{kifu} object}

\item{y}{not in use (just for argument consistency with generic function)}

\item{...}{arguments passed to \code{\link{ggoban}}}
}
\value{
\code{ggplot} object
}
\description{
Draw kifu
}

