% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gogame-class.R
\name{plotat}
\alias{plotat}
\title{Plot the go board state by ggplot}
\usage{
plotat(x, at, marklast = TRUE, lastmarker = utf8ToInt(9650), ...)
}
\arguments{
\item{x}{\code{gogame} object}

\item{at}{Move number (integer)}

\item{marklast}{If specifified, add a marker to the last stone (logical)}

\item{lastmarker}{character of marker indicating the last move}

\item{...}{graphic parameters}
}
\value{
\code{ggplot} object
}
\description{
Plot the go board state by ggplot
}

