% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gogame-class.R
\name{plotat}
\alias{plotat}
\title{Plot the go board state by ggplot}
\usage{
plotat(x, at, stonesize = 6, marklastmove = "", blackcolor = "#000000",
  whitecolor = "#ffffff", linecolor = "#000000", ...)
}
\arguments{
\item{x}{\code{gogame} object}

\item{at}{integer of the move number}

\item{stonesize}{numeric that indicates the size of stones,
passed to \code{geom_point}}

\item{blackcolor}{color for black stone}

\item{whitecolor}{color for white stone}

\item{linecolor}{color for stone outline}

\item{...}{arguments passed to \code{\link{ggoban}}}
}
\value{
\code{ggplot} object
}
\description{
Plot the go board state by ggplot
}

