% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gokifu-class.R
\name{print.gokifu}
\alias{print.gokifu}
\title{Print kifu on console}
\usage{
\method{print}{gokifu}(x, ...)
}
\arguments{
\item{x}{\code{gokifu} object}

\item{...}{graphical parameters}
}
\description{
Print kifu on console
}

