% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{prune_sgf}
\alias{prune_sgf}
\title{Remove branches from SGF text}
\usage{
prune_sgf(sgf, keepfirst = TRUE)
}
\arguments{
\item{sgf}{Scalar character formatted as SGF}

\item{keepfirst}{Logical. If TRUE, keep the first branch.
Otherwise, keep the last branch.}
}
\value{
Scalar character of SGF text where branches are removed
}
\description{
Remove branches from SGF text
}

