% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prune.R
\name{prune_sgf}
\alias{prune_sgf}
\title{Remove brances of SGF}
\usage{
prune_sgf(sgf)
}
\arguments{
\item{sgf}{Scalar character formatted as SGF}
}
\value{
Scalar character of SGF where branches are removed
}
\description{
Remove brances of SGF
}

