% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reader.R
\name{read_sgf}
\alias{read_sgf}
\title{Read and parse a SGF file}
\usage{
read_sgf(path, keep_first = TRUE, encoding = "")
}
\arguments{
\item{path}{A character string of the path to
a smart go format (SGF) file.  Can be local or online.}

\item{encoding}{A character string.
If non-empty declares the encoding used on a file.}
}
\description{
Read and parse a SGF file
}

