% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sgfreader.R
\name{read_sgf}
\alias{read_sgf}
\title{Read and parse a SGF file}
\usage{
read_sgf(path, ...)
}
\arguments{
\item{path}{character string of the path to
a smart go format (SGF) file. Can be local or online}

\item{...}{arguments passed to \code{\link{readLines}}}
}
\value{
\code{\link{gogame}} object
}
\description{
Read and parse a SGF file
}

