% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reader.R
\name{read_sgf}
\alias{read_sgf}
\title{Read and parse a SGF file}
\usage{
read_sgf(path, keepfirst = TRUE, ...)
}
\arguments{
\item{path}{character string of the path to
a smart go format (SGF) file. Can be local or online}

\item{keepfirst}{logical specifying the branch choice rule.
If TRUE, keep the first appearing branch.
Otherwise, keep the last branch.}

\item{...}{arguments passed to \code{\link{readLines}}}
}
\value{
\code{\link{gogame}} object
}
\description{
Read and parse a SGF file
}

