% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{saikoyo}
\alias{saikoyo}
\title{sai vs toya koyo}
\format{\code{gogame} object}
\source{
Hotta, Yumi, Obata, Takeshi, and Umezawa, Yukari. Hikaru no Go: Gojasu Characters Guide. Shueisha 2002.
}
\usage{
saikoyo
}
\description{
Go game between Sai and Toya Koyo played online
in the comic "Hikaru no Go"
}
\details{
The game played between Sai and Toya Koyo on internet in the comic
"Hikaru no Go", Episode 112-115.
The game is originally played by Japanese professionals
Yoda Notimoto and Rin Kaiho on May 1st, 1997.
In the comic, Toya Koyo resigns when he realizes that
the game would eventually end with his loss by 0.5 point.
}
\keyword{datasets}

