% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{saikoyo}
\alias{saikoyo}
\title{sai vs toya koyo}
\format{\code{gogame} object}
\source{
Hotta, Yumi, Obata, Takeshi, and Umezawa, Yukari. Hikaru no Go: Gojasu Characters Guide. Shueisha 2002.
}
\usage{
saikoyo
}
\description{
Go game between sai and Toya Koyo played online
in the comic "Hikaru no Go"
}
\details{
The game played between Sai and Toya Koyo on internet in the comic
"Hikaru no Go", Episode 112-115.
The game is identical to a Japanese professional game between
Yoda Notimoto and Rin Kaiho on May 1st, 1997.
}
\keyword{datasets}

