% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gogame-class.R
\name{set_gamepath}
\alias{set_gamepath}
\title{Switch path of go game}
\usage{
set_gamepath(x, pathid = 1L)
}
\arguments{
\item{x}{\code{gogame} object}

\item{pathid}{integer}
}
\value{
\code{gogame} object
}
\description{
Switch path of a go game. Paths are indexed by integers starting
at one. If pathid exceeds the number of paths stored in the game, the function
throws an error.
}

