% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gogame-class.R
\name{stateat}
\alias{stateat}
\title{Go board status at a move number}
\usage{
stateat(x, at)
}
\arguments{
\item{x}{\code{gogame} object}

\item{at}{integer of the move number}
}
\value{
\code{\link{gostate}} object
}
\description{
This function obtains the board state at the move number.
The result is stored in a \code{\link{gostate}} object.
}
\examples{
stateat(saikoyo, 116)
}

