% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gogame-class.R
\name{stateat}
\alias{plotat}
\alias{stateat}
\title{Go board status at a move number}
\usage{
stateat(x, at)

plotat(x, at, ...)
}
\arguments{
\item{x}{\code{gogame} object}

\item{at}{integer of the move number}

\item{...}{arguments passed to \code{\link{plot.gostate}}}
}
\value{
\code{stateat} returns a \code{\link{gostate}} object

\code{plotat} returns a \code{ggplot} object
}
\description{
Computes the board state at a certain move number for a game.
}
\examples{
stateat(saikoyo, 116)
plotat(mimiaka, 127)
}

